Split out IBuffer into a separate file
